requirements.txt
selenium-wire
ipdb
requests
Flask-SQLAlchemy
humanize